worker/requirements.txt
redis==5.0.8
rq==1.16.2
pypdf==4.0.0
pdf2docx==0.5.8
python-docx==1.1.2
reportlab==4.2.0
pydantic==2.9.0
